Removing unnecessary code

+++ src/main/java/org/javamoney/moneta/convert/internal/ECBHistoricRateProvider.java
@@ -142,7 +142,7 @@ public class ECBHistoricRateProvider extends AbstractRateProvider implements Loa
         cal.set(Calendar.MINUTE, 0);
         cal.set(Calendar.SECOND, 0);
         cal.set(Calendar.MILLISECOND, 0);
-        Long targetTS = Long.valueOf(cal.getTimeInMillis());
+        Long targetTS = cal.getTimeInMillis();
         Map<String,ExchangeRate> targets = this.historicRates.get(targetTS);
         if(targets == null){
             return null;
@@ -233,7 +233,7 @@ public class ECBHistoricRateProvider extends AbstractRateProvider implements Loa
                 if("Cube".equals(qName)){
                     if(attributes.getValue("time") != null){
                         Date date = dateFormat.parse(attributes.getValue("time"));
-                        timestamp = Long.valueOf(date.getTime());
+                        timestamp = date.getTime();
                     }else if(attributes.getValue("currency") != null){
                         // read data <Cube currency="USD" rate="1.3349"/>
                         CurrencyUnit tgtCurrency = MonetaryCurrencies.getCurrency(attributes.getValue("currency"));
@@ -261,7 +261,7 @@ public class ECBHistoricRateProvider extends AbstractRateProvider implements Loa
         RateType rateType = RateType.HISTORIC;
         ExchangeRate.Builder builder = null;
         if(timestamp != null){
-            if(timestamp.longValue() > System.currentTimeMillis()){
+            if(timestamp > System.currentTimeMillis()){
                 rateType = RateType.DEFERRED;
             }
             builder = new ExchangeRate.Builder(new ConversionContext.Builder(CONTEXT, rateType).setAttribute(TIMESTAMP, timestamp).build());
